requirements.txt
discord.py
feedparser
python-dotenv
yfinance
pandas
pytz
requests
numpy
joblib
scikit-learn




